+++ modules/face/__init__.py
@@ -105,14 +105,13 @@ class Script(scripts.Script):
         for i, image in enumerate(input_images):
             if not isinstance(image, Image.Image):
                 input_images[i] = Image.open(image['name'])
-        source_image = input_images[0]
 
         processing.process_init(p)
         if mode == 'FaceID': # faceid runs as ipadapter in its own pipeline
             from modules.face.insightface import get_app
             app = get_app('buffalo_l')
             from modules.face.faceid import face_id
-            processed_images = face_id(p, app=app, source_image=source_image, model=ip_model, override=ip_override, cache=ip_cache, scale=ip_strength, structure=ip_structure) # run faceid pipeline
+            processed_images = face_id(p, app=app, source_images=input_images, model=ip_model, override=ip_override, cache=ip_cache, scale=ip_strength, structure=ip_structure) # run faceid pipeline
             processed = processing.Processed(p, images_list=processed_images, seed=p.seed, subseed=p.subseed, index_of_first_image=0) # manually created processed object
         elif mode == 'PhotoMaker': # photomaker creates pipeline and triggers original process_images
             from modules.face.photomaker import photo_maker
@@ -121,7 +120,7 @@ class Script(scripts.Script):
             from modules.face.insightface import get_app
             app=get_app('antelopev2')
             from modules.face.instantid import instant_id # instantid creates pipeline and triggers original process_images
-            processed = instant_id(p, app=app, source_image=source_image, strength=id_strength, conditioning=id_conditioning, cache=id_cache)
+            processed = instant_id(p, app=app, source_images=input_images, strength=id_strength, conditioning=id_conditioning, cache=id_cache)
 
         if processed is None: # run normal pipeline
             processed = processing.process_images(p)
@@ -134,7 +133,7 @@ class Script(scripts.Script):
                 for i, image in enumerate(processed.images):
                     info = processing.create_infotext(p, index=i)
                     images.save_image(image, path=p.outpath_samples, seed=p.all_seeds[i], prompt=p.all_prompts[i], info=info, p=p, suffix="-before-faceswap")
-            processed.images = face_swap(p, app=app, input_images=processed.images, source_image=source_image, cache=fs_cache)
+            processed.images = face_swap(p, app=app, input_images=processed.images, source_image=input_images[0], cache=fs_cache)
 
         processed.info = processed.infotext(p, 0)
         processed.infotexts = [processed.info]