+++ modules/face/insightface.py
@@ -1,5 +1,6 @@
 import os
 from modules.shared import log, opts
+from modules import devices
 
 
 insightface_app = None
@@ -17,11 +18,10 @@ def get_app(mp_name):
             install(pkg[0], pkg[1], ignore=True)
     global insightface_app, instightface_mp # pylint: disable=global-statement
     if insightface_app is None or mp_name != instightface_mp:
-        import onnxruntime
         from insightface.app import FaceAnalysis
         import huggingface_hub as hf
         import zipfile
-        log.debug(f"InsightFace: mp={mp_name} device={onnxruntime.get_device()} providers={onnxruntime.get_available_providers()}")
+        log.debug(f"InsightFace: mp={mp_name} provider={devices.onnx}")
         root_dir = os.path.join(opts.diffusers_dir, 'models--vladmandic--insightface-faceanalysis')
         local_dir = os.path.join(root_dir, 'models')
         extract_dir = os.path.join(local_dir, mp_name)
@@ -44,8 +44,7 @@ def get_app(mp_name):
             'download': False,
             'download_zip': False,
         }
-        insightface_app = FaceAnalysis(name=mp_name, providers=['CUDAExecutionProvider', 'CPUExecutionProvider'], **kwargs)
+        insightface_app = FaceAnalysis(name=mp_name, providers=devices.onnx, **kwargs)
         instightface_mp = mp_name
-        onnxruntime.set_default_logger_severity(3)
         insightface_app.prepare(ctx_id=0, det_thresh=0.5, det_size=(640, 640))
     return insightface_app