+++ modules/loader.py
@@ -37,6 +37,10 @@ if ".dev" in torch.__version__ or "+git" in torch.__version__:
     torch.__version__ = re.search(r'[\d.]+[\d]', torch.__version__).group(0)
 timer.startup.record("torch")
 
+import onnxruntime
+onnxruntime.set_default_logger_severity(3)
+timer.startup.record("onnx")
+
 from modules.onnx_impl import initialize_olive # pylint: disable=ungrouped-imports
 initialize_olive()
 timer.startup.record("olive")