+++ modules/onnx_impl/execution_providers.py
@@ -33,7 +33,6 @@ TORCH_DEVICE_TO_EP = {
 
 def get_default_execution_provider() -> ExecutionProvider:
     from modules import devices
-
     if devices.backend == "cpu":
         return ExecutionProvider.CPU
     elif devices.backend == "directml":
@@ -41,10 +40,7 @@ def get_default_execution_provider() -> ExecutionProvider:
     elif devices.backend == "cuda":
         return ExecutionProvider.CUDA
     elif devices.backend == "rocm":
-        if ExecutionProvider.ROCm in available_execution_providers:
-            return ExecutionProvider.ROCm
-        else:
-            return ExecutionProvider.CPU
+        return ExecutionProvider.ROCm
     elif devices.backend == "ipex" or devices.backend == "openvino":
         return ExecutionProvider.OpenVINO
     return ExecutionProvider.CPU
@@ -52,11 +48,7 @@ def get_default_execution_provider() -> ExecutionProvider:
 
 def get_execution_provider_options():
     from modules.shared import cmd_opts, opts
-
-    execution_provider_options = {
-        "device_id": int(cmd_opts.device_id or 0),
-    }
-
+    execution_provider_options = { "device_id": int(cmd_opts.device_id or 0) }
     if opts.onnx_execution_provider == ExecutionProvider.ROCm:
         if ExecutionProvider.ROCm in available_execution_providers:
             execution_provider_options["tunable_op_enable"] = 1
@@ -68,32 +60,21 @@ def get_execution_provider_options():
             raw_openvino_device = f"{raw_openvino_device}_FP16"
         execution_provider_options["device_type"] = raw_openvino_device
         del execution_provider_options["device_id"]
-
     return execution_provider_options
 
 
 def get_provider() -> Tuple:
     from modules.shared import opts
-
     return (opts.onnx_execution_provider, get_execution_provider_options(),)
 
 
 def install_execution_provider(ep: ExecutionProvider):
-    from installer import pip, uninstall, installed, get_onnxruntime_source_for_rocm
-
-    if installed("onnxruntime"):
-        uninstall("onnxruntime")
-    if installed("onnxruntime-directml"):
-        uninstall("onnxruntime-directml")
-    if installed("onnxruntime-gpu"):
-        uninstall("onnxruntime-gpu")
-    if installed("onnxruntime-training"):
-        uninstall("onnxruntime-training")
-    if installed("onnxruntime-openvino"):
-        uninstall("onnxruntime-openvino")
-
+    import imp  # pylint: disable=deprecated-module
+    from installer import installed, install, uninstall, get_onnxruntime_source_for_rocm
+    res = "<br><pre>"
+    res += uninstall(["onnxruntime", "onnxruntime-directml", "onnxruntime-gpu", "onnxruntime-training", "onnxruntime-openvino"], quiet=True)
+    installed("onnxruntime", reload=True)
     packages = ["onnxruntime"] # Failed to load olive: cannot import name '__version__' from 'onnxruntime'
-
     if ep == ExecutionProvider.DirectML:
         packages.append("onnxruntime-directml")
     elif ep == ExecutionProvider.CUDA:
@@ -102,13 +83,14 @@ def install_execution_provider(ep: ExecutionProvider):
         if "linux" not in sys.platform:
             log.warning("ROCMExecutionProvider is not supported on Windows.")
             return
-
         packages.append(get_onnxruntime_source_for_rocm(None))
     elif ep == ExecutionProvider.OpenVINO:
-        if installed("openvino"):
-            uninstall("openvino")
         packages.append("openvino")
         packages.append("onnxruntime-openvino")
-
-    pip(f"install --upgrade {' '.join(packages)}")
-    log.info("Please restart SD.Next.")
+    for package in packages:
+        res += install(package)
+    res += '</pre><br>'
+    res += 'Server restart required'
+    log.info("Server restart required")
+    imp.reload(ort)
+    return res